Deleted the accountcreation file from router, as it was not needed. Fixed formatting issues in the account.js
