--- server/routes/publicAccountCreation.js
@@ -1,19 +1,0 @@
-const router = express.Router();
-
- const accountHandler = require('../models/handlers/account');
-
- router.get('/api/create-user', async(req, res) => {
-    const firstName = req.query.firstName;
-    const lastName = req.query.lastName;
-    const type = req.query.type; //note: 'type' is called 'role' in the UI, but refers to either a worker or a student.
-    const studentID = req.query.studentID;
-    const email = req.query.email;
-    const password = req.query.password;
-    const phone = req.query.phone;
-
-     const user = await accountHandler.createUserAccount(firstName, lastName, type, studentID, email, password, phone);
-
-     res.send(user);
- });
-
- module.exports = router
